# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PeopleMatrix synthetic HR lab API. Assessment details are not published here.
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type AccountRole = typeof AccountRole[keyof typeof AccountRole];


export const AccountRole = {
  EMPLOYEE: 'EMPLOYEE',
  MANAGER: 'MANAGER',
  HR_ADMIN: 'HR_ADMIN',
  SYSTEM_ADMIN: 'SYSTEM_ADMIN',
} as const;

export interface Account {
  id: number;
  email: string;
  role: AccountRole;
  /** @nullable */
  employeeId: number | null;
  active: boolean;
}

export interface LoginResult {
  token: string;
  user: Account;
}

export interface EmployeeSummary {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  jobTitle: string;
  department: string;
  /** @nullable */
  managerId: number | null;
}

export type EmployeeProfile = EmployeeSummary & {
  phone: string;
  employmentDate: string;
  employmentStatus: string;
  address: string;
  emergencyContact: string;
  leaveBalance: number;
};

export interface EmployeeInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  phone: string;
  jobTitle: string;
  department: string;
  /** @nullable */
  managerId?: number | null;
  employmentDate: string;
  salary: string;
  bankName: string;
  bankAccount: string;
  syntheticNin: string;
  address: string;
  emergencyContact: string;
}

export interface EmployeeUpdate {
  phone?: string;
  address?: string;
  emergencyContact?: string;
  jobTitle?: string;
  department?: string;
  /** @nullable */
  managerId?: number | null;
  employmentStatus?: string;
  salary?: string;
  bankName?: string;
  bankAccount?: string;
  syntheticNin?: string;
}

export interface PayrollRecord {
  id: number;
  employeeId: number;
  period: string;
  gross: string;
  deductions: string;
  net: string;
  paymentDate: string;
}

export interface LeaveInput {
  type: string;
  startsOn: string;
  endsOn: string;
  reason: string;
}

export type LeaveDecisionStatus = typeof LeaveDecisionStatus[keyof typeof LeaveDecisionStatus];


export const LeaveDecisionStatus = {
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
} as const;

export interface LeaveDecision {
  status: LeaveDecisionStatus;
}

export interface LeaveRecord {
  id: number;
  employeeId: number;
  type: string;
  startsOn: string;
  endsOn: string;
  reason: string;
  status: string;
  /** @nullable */
  reviewerId: number | null;
  createdAt: string;
}

export interface PerformanceInput {
  employeeId: number;
  period: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
}

export interface PerformanceRecord {
  id: number;
  employeeId: number;
  reviewerId: number;
  period: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export type ReportSummaryDepartmentCounts = {[key: string]: number};

export interface ReportSummary {
  employeeCount: number;
  departmentCounts: ReportSummaryDepartmentCounts;
  pendingLeave: number;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  EMPLOYEE: 'EMPLOYEE',
  MANAGER: 'MANAGER',
  HR_ADMIN: 'HR_ADMIN',
  SYSTEM_ADMIN: 'SYSTEM_ADMIN',
} as const;

export interface UserInput {
  email: string;
  /** @minLength 12 */
  password: string;
  role: UserInputRole;
  /** @nullable */
  employeeId?: number | null;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  EMPLOYEE: 'EMPLOYEE',
  MANAGER: 'MANAGER',
  HR_ADMIN: 'HR_ADMIN',
  SYSTEM_ADMIN: 'SYSTEM_ADMIN',
} as const;

export interface UserUpdate {
  role?: UserUpdateRole;
  active?: boolean;
  /** @minLength 12 */
  password?: string;
}

export interface Setting {
  key: string;
  value: string;
}

export interface AuditEvent {
  id: number;
  createdAt: string;
  /** @nullable */
  userId: number | null;
  /** @nullable */
  ip: string | null;
  method: string;
  endpoint: string;
  action: string;
  status: number;
}

export type ListEmployeesParams = {
q?: string;
department?: string;
};

export type SearchEmployeesParams = {
department: string;
};

export type ListPerformanceParams = {
employeeId?: number;
};