# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PeopleMatrix synthetic HR lab API. Assessment details are not published here.
 * OpenAPI spec version: 0.2.0
 */
import type { AccountRole } from './accountRole';

export interface Account {
  id: number;
  email: string;
  role: AccountRole;
  /** @nullable */
  employeeId: number | null;
  active: boolean;
}