# lib/api-zod/src/generated/types/auditEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PeopleMatrix synthetic HR lab API. Assessment details are not published here.
 * OpenAPI spec version: 0.2.0
 */

export interface AuditEvent {
  id: number;
  createdAt: string;
  /** @nullable */
  userId: number | null;
  /** @nullable */
  ip: string | null;
  method: string;
  endpoint: string;
  action: string;
  status: number;
}