# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PeopleMatrix synthetic HR lab API. Assessment details are not published here.
 * OpenAPI spec version: 0.2.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  role?: UserUpdateRole;
  active?: boolean;
  /** @minLength 12 */
  password?: string;
}